+++ ark_auth/src/client/async_impl/user.rs
@@ -1,16 +1,17 @@
-use crate::client::{Client, ClientError};
+use crate::client::async_impl::AsyncClient;
+use crate::client::Error;
 use crate::server::route::user::{CreateBody, CreateResponse};
 use actix_web::http::StatusCode;
 use futures::{future, Future};
 
-impl Client {
+impl AsyncClient {
     pub fn user_create(
         &self,
         name: &str,
         email: &str,
         active: bool,
         password: Option<&str>,
-    ) -> impl Future<Item = CreateResponse, Error = ClientError> {
+    ) -> impl Future<Item = CreateResponse, Error = Error> {
         let body = CreateBody {
             name: name.to_owned(),
             email: email.to_owned(),
@@ -20,14 +21,11 @@ impl Client {
 
         self.post("/v1/user")
             .send_json(&body)
-            .map_err(|_err| ClientError::Unwrap)
+            .map_err(|_err| Error::Unwrap)
             .and_then(|res| match res.status() {
                 StatusCode::OK => future::ok(res),
-                _ => future::err(ClientError::Unwrap),
+                _ => future::err(Error::Unwrap),
             })
-            .and_then(|mut res| {
-                res.json::<CreateResponse>()
-                    .map_err(|_err| ClientError::Unwrap)
-            })
+            .and_then(|mut res| res.json::<CreateResponse>().map_err(|_err| Error::Unwrap))
     }
 }